pyproject.toml
[tool.poetry]
name = "vault-rummager"
version = "0.1.0"
description = "A private, local AI chatbot that rummages through your Obsidian vault."
authors = ["Lindsey Burnett <lburnett@redhat.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sentence-transformers = "^4.1.0"
transformers = "^4.51.3"
torch = "^2.7.0"
python-frontmatter = "^1.1.0"
chromadb = {extras = ["rest"], version = "^1.0.7"}
nltk = "^3.9.1"
rich = "^14.0.0"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
